added custom convex glb collisions for links 2,3,4

+++ mani_skill/envs/tasks/digital_twins/koch_arm_tabletop/transfer_cube.py
@@ -158,7 +158,7 @@ class TransferCubeEnv(BaseEnv):
             "success": is_obj_placed & ~is_grasped & (robot_to_rest_pose_dist < 0.15),
             "is_obj_placed": is_obj_placed,
             "within_end_boundary_xy": within_end_boundary_xy,
-            "robot_to_rest_pose_dist": robot_to_rest_pose_dist,
+            "robot_to_rest_pose_dist": robot_to_rest_pose_dist.float(),
             "is_grasped": is_grasped,
         }
